+++ src/pysiral/cryosat2/iotools.py
@@ -9,7 +9,7 @@ import numpy as np
 from loguru import logger
 
 from pysiral.core import DefaultLoggingClass
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import ErrorStatus
 
 
 class CryoSat2MonthlyFileListAllModes(DefaultLoggingClass):