+++ src/pysiral/envisat/iotools.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import List
 
 from pysiral.core import DefaultLoggingClass
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import ErrorStatus
 
 
 class EnvisatSGDRNC(DefaultLoggingClass):