+++ src/pysiral/l1data.py
@@ -100,7 +100,7 @@ from loguru import logger
 from netCDF4 import Dataset, date2num
 from scipy.spatial.transform import Rotation
 
-from pysiral.core.class_template import DefaultLoggingClass
+from pysiral.core.legacy_classes import DefaultLoggingClass
 from pysiral.core.config import RadarModes
 from pysiral.core.flags import SurfaceType
 from pysiral.core.output import NCDateNumDef