+++ src/pysiral/mask.py
@@ -21,8 +21,8 @@ from netCDF4 import Dataset
 from pyresample import geometry, image, kd_tree
 
 from pysiral import psrlcfg
-from pysiral.core.class_template import DefaultLoggingClass
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import DefaultLoggingClass
+from pysiral.core.legacy_classes import ErrorStatus
 from pysiral.core.flags import SURFACE_TYPE_DICT
 from pysiral.core.iotools import ReadNC
 from pysiral.grid import GridDefinition, GridTrajectoryExtract